refactor(deploy): restructure the deploy command around plan/deploy adapters
