+++ packages/@sanity/cli/src/actions/deploy/__tests__/runDeploy.test.ts
@@ -2,18 +2,22 @@ import {type Dirent, type Stats} from 'node:fs'
 import {readdir, stat} from 'node:fs/promises'
 import {join} from 'node:path'
 
+import {CLIError} from '@oclif/core/errors'
 import {type Output} from '@sanity/cli-core'
 import {beforeEach, describe, expect, test, vi} from 'vitest'
 
-import {createCollectingReporter, type DeployCheck} from '../deployChecks.js'
+import {type DeployCheck} from '../checks.js'
 import {
+  type DeployAdapter,
   type DeploymentFile,
   type DeploymentPlan,
   deploymentPlanToJson,
   listDeploymentFiles,
+  newPlan,
   renderDeploymentPlan,
-  reportExposes,
-} from '../deploymentPlan.js'
+  runDeploy,
+} from '../runDeploy.js'
+import {type DeployAppOptions} from '../types.js'
 
 vi.mock(import('node:fs/promises'), async (importOriginal) => ({
   ...(await importOriginal()),
@@ -56,15 +60,11 @@ describe('listDeploymentFiles', () => {
   })
 })
 
-const studioPlan = (checks: DeployCheck[], files: DeploymentFile[] = []): DeploymentPlan => ({
-  checks,
-  exposes: [],
-  files,
-  installationConfig: null,
-  target: null,
-  type: 'studio',
-  version: '3.99.0',
-})
+const studioPlan = (
+  checks: DeployCheck[],
+  files: DeploymentFile[] = [],
+): DeploymentPlan<'studio'> =>
+  newPlan({checks, files, target: null, type: 'studio', version: '3.99.0'})
 
 describe('deploymentPlanToJson', () => {
   test('maps failing checks to fixes and warnings to messages, dropping pass/skip', () => {
@@ -98,15 +98,7 @@ describe('deploymentPlanToJson', () => {
       title: 'My Studio',
       url: 'https://my-studio.sanity.studio',
     }
-    const json = deploymentPlanToJson({
-      checks: [],
-      exposes: [],
-      files: [],
-      installationConfig: null,
-      target,
-      type: 'studio',
-      version: null,
-    })
+    const json = deploymentPlanToJson(newPlan({checks: [], target, type: 'studio', version: null}))
     expect(json.target).toEqual(target)
   })
 
@@ -131,31 +123,6 @@ describe('deploymentPlanToJson', () => {
 
     expect(json.exposes).toEqual([{name: 'edit', title: 'Edit', type: 'panel'}])
     expect(json.installationConfig).toBe('Media Library fields:\n  Title (title)')
-  })
-})
-
-describe('reportExposes', () => {
-  test('reports views and services and returns them structured', () => {
-    const reporter = createCollectingReporter()
-
-    const exposes = reportExposes(reporter, {
-      services: [{name: 'sync', src: './sync.ts', type: 'worker'}],
-      views: [{name: 'edit', src: './edit.ts', title: 'Edit', type: 'panel'}],
-    })
-
-    expect(exposes).toEqual([
-      {name: 'edit', title: 'Edit', type: 'panel'},
-      {name: 'sync', title: 'sync', type: 'worker'},
-    ])
-    expect(reporter.results.every((check) => check.status === 'pass')).toBe(true)
-    // The structured list rides on the first check so a dry run's collector reads it back.
-    expect(reporter.results[0].exposes).toEqual(exposes)
-  })
-
-  test('reports nothing and returns empty without views or services', () => {
-    const reporter = createCollectingReporter()
-    expect(reportExposes(reporter, {})).toEqual([])
-    expect(reporter.results).toEqual([])
   })
 })
 
@@ -229,18 +196,160 @@ describe('renderDeploymentPlan', () => {
 
   test('labels a core app deploy as an application', () => {
     renderDeploymentPlan(
-      {
-        checks: [],
-        exposes: [],
-        files: [],
-        installationConfig: null,
-        target: null,
-        type: 'coreApp',
-        version: '1.0.0',
-      },
+      newPlan({checks: [], target: null, type: 'coreApp', version: '1.0.0'}),
       output,
     )
 
     expect(lines.join('\n')).toContain('This application can be deployed.')
+  })
+})
+
+const mockOutput = () => ({error: vi.fn(), log: vi.fn(), warn: vi.fn()}) as unknown as Output
+
+const dryRunOptions = (output: Output): DeployAppOptions =>
+  ({
+    cliConfig: {},
+    flags: {'dry-run': true},
+    output,
+    projectRoot: {directory: '/root', path: '/root/sanity.cli.ts'},
+    sourceDir: '/root/dist',
+  }) as unknown as DeployAppOptions
+
+const studioAdapter = (
+  overrides: Partial<DeployAdapter<'studio'>> = {},
+): DeployAdapter<'studio'> => ({
+  deploy: async () => undefined,
+  plan: async () => newPlan({checks: [], target: null, type: 'studio', version: null}),
+  type: 'studio',
+  ...overrides,
+})
+
+describe('runDeploy dry run', () => {
+  test('exits with the first failing check exit code, like a real deploy', async () => {
+    const output = mockOutput()
+    const adapter = studioAdapter({
+      plan: async () => studioPlan([{exitCode: 2, message: 'boom', status: 'fail'}]),
+    })
+
+    await runDeploy(dryRunOptions(output), adapter)
+
+    expect(output.error).toHaveBeenCalledWith('Deploy blocked by failing checks.', {exit: 2})
+  })
+
+  test('a deployable dry run renders the plan without erroring', async () => {
+    const output = mockOutput()
+    const adapter = studioAdapter({
+      plan: async () =>
+        studioPlan(
+          [{message: 'Project: p1', status: 'pass'}],
+          [{path: 'dist/index.html', size: 10}],
+        ),
+    })
+
+    await runDeploy(dryRunOptions(output), adapter)
+
+    expect(output.error).not.toHaveBeenCalled()
+    expect(output.log).toHaveBeenCalledWith(expect.stringContaining('This studio can be deployed.'))
+  })
+
+  test('the JSON plan reports the version the plan resolved', async () => {
+    const output = mockOutput()
+    const adapter = studioAdapter({
+      plan: async () =>
+        newPlan({
+          checks: [{message: 'Using sanity 9.9.9', status: 'pass'}],
+          target: null,
+          type: 'studio',
+          version: '9.9.9',
+        }),
+    })
+
+    await runDeploy(
+      {...dryRunOptions(output), flags: {'dry-run': true, json: true}} as DeployAppOptions,
+      adapter,
+    )
+
+    const payload = JSON.parse(vi.mocked(output.log).mock.calls.at(-1)![0] as string)
+    expect(payload.applicationVersion).toBe('9.9.9')
+  })
+
+  test('never calls deploy', async () => {
+    const output = mockOutput()
+    const deploy = vi.fn()
+    const adapter = studioAdapter({deploy})
+
+    await runDeploy(dryRunOptions(output), adapter)
+
+    expect(deploy).not.toHaveBeenCalled()
+  })
+
+  test('a blocked --json dry run prints the plan only, never a deploy envelope', async () => {
+    const output = mockOutput()
+    // A real run's output.error throws to abort; the plan JSON is already out.
+    vi.mocked(output.error).mockImplementation(() => {
+      throw new CLIError('blocked')
+    })
+    const adapter = studioAdapter({
+      plan: async () => studioPlan([{exitCode: 2, message: 'boom', status: 'fail'}]),
+    })
+
+    await expect(
+      runDeploy(
+        {...dryRunOptions(output), flags: {'dry-run': true, json: true}} as DeployAppOptions,
+        adapter,
+      ),
+    ).rejects.toThrow()
+
+    const logged = vi.mocked(output.log).mock.calls.map((call) => call[0] as string)
+    expect(logged).toHaveLength(1)
+    expect(logged[0]).not.toContain('"deployed"')
+  })
+})
+
+describe('runDeploy real deploy', () => {
+  test('emits the deploy result as JSON, marked deployed', async () => {
+    const output = mockOutput()
+    const result = {
+      applicationType: 'studio' as const,
+      applicationVersion: '3.99.0',
+      target: {applicationId: 'app-1', title: 'My Studio', url: 'https://my-studio.sanity.studio'},
+    }
+    const adapter = studioAdapter({deploy: async () => result})
+
+    await runDeploy({...dryRunOptions(output), flags: {json: true}} as DeployAppOptions, adapter)
+
+    const payload = JSON.parse(vi.mocked(output.log).mock.calls.at(-1)![0] as string)
+    expect(payload).toEqual({deployed: true, ...result})
+  })
+
+  test('a failed deploy emits a {deployed: false} envelope and still errors on stderr', async () => {
+    const output = mockOutput()
+    const adapter = studioAdapter({
+      deploy: async () => {
+        throw new Error('boom')
+      },
+    })
+
+    await runDeploy({...dryRunOptions(output), flags: {json: true}} as DeployAppOptions, adapter)
+
+    const payload = JSON.parse(vi.mocked(output.log).mock.calls.at(-1)![0] as string)
+    expect(payload.deployed).toBe(false)
+    expect(payload.error.message).toContain('boom')
+    // The envelope and the stderr message are the same diagnosis
+    expect(output.error).toHaveBeenCalledWith(payload.error.message, {exit: 1})
+  })
+
+  test('without --json a failed deploy stays on stderr, with no envelope', async () => {
+    const output = mockOutput()
+    const adapter = studioAdapter({
+      deploy: async () => {
+        throw new Error('boom')
+      },
+    })
+
+    await runDeploy({...dryRunOptions(output), flags: {}} as DeployAppOptions, adapter)
+
+    expect(output.log).not.toHaveBeenCalled()
+    expect(output.error).toHaveBeenCalledWith(expect.stringContaining('boom'), {exit: 1})
   })
 })
